+++ client/src/api/client.ts
@@ -3,7 +3,7 @@ import useInterceptors from './interceptors';
 
 const api = axios.create({
     baseURL: `${import.meta.env.VITE_API_BASE_URL}/api`,
-    withCredentials: true
+    withCredentials: true,
 });
 
 export default useInterceptors(api);