+++ client/src/api/index.ts
@@ -1,2 +1,1 @@
 export { default as client } from './client';
-export { default as endpoints } from './endpoints';