+++ client/src/api/interceptors/index.ts
@@ -2,7 +2,7 @@ import type { AxiosInstance } from 'axios';
 import useAuthorization from './useAuthorization';
 import useRefreshToken from './useRefreshToken';
 
-export default function(api: AxiosInstance) {
+export default function (api: AxiosInstance) {
     useAuthorization(api);
     useRefreshToken(api);
     return api;