+++ client/src/config/index.ts
@@ -1,7 +1,9 @@
 import routes from './routes';
+import endpoints from './endpoints';
 
 const config = {
-    routes
-}
+    routes,
+    endpoints,
+};
 
 export default config;