+++ client/src/core/index.tsx
@@ -7,14 +7,11 @@ export const AccessorieController = new MainResourceController(
     Services.AccessorieService,
 );
 export const AssetController = new MainResourceController(
-    config.routes.assets, 
-    Services.AssetService
+    config.routes.assets,
+    Services.AssetService,
 );
 export const LicenseController = new MainResourceController(
     config.routes.licenses,
     Services.LicenseService,
 );
-export const UserController = new MainResourceController(
-    config.routes.users, 
-    Services.UserService
-);
+export const UserController = new MainResourceController(config.routes.users, Services.UserService);