+++ client/src/types/index.ts
@@ -1,11 +1,6 @@
 export type { default as ILoginForm } from './forms/LoginForm';
 export type { default as ILoginResponse } from './responses/LoginResponse';
-
 export type { default as IAuthContextType } from './contexts/AuthContextType';
-
 export type { default as IUser } from './resources/user';
-export type { default as IAccessorie } from './resources/accessorie';
-export type { default as IAsset } from './resources/asset';
-export type { default as ILicense } from './resources/license';
 export type { default as IHistoryRecord } from './resources/historyRecord';
 export type { IResource, IMainResource } from './services/service';