+++ common/helpers/getLastDocument.js
@@ -1,6 +1,6 @@
 import ConstMessages from '../constants/constMessages.js';
 
-const getLastDocument = async(model) => {
+const getLastDocument = async model => {
     try {
         const lastUser = await model.findOne().sort({ id: -1 });
         return lastUser ? lastUser.id + 1 : 1;